feat: refactor authentication routes and services to use environment variables
